preparing checker for first argument

+++ client.c
@@ -6,22 +6,38 @@
 /*   By: gforns-s <gforns-s@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 13:46:01 by gforns-s          #+#    #+#             */
-/*   Updated: 2023/09/06 17:59:56 by gforns-s         ###   ########.fr       */
+/*   Updated: 2023/09/06 18:38:39 by gforns-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minitalk.h"\
 
+int return_error(void)
+{
+	ft_printf("Error/n");
+	return (-1);
+}
+
+
+int	check_if_digit_only(int argc, char *argv)
+{
+
+
+
+}
+
+
 int	main(int argc, char **argv)
 {
 	pid_t	pid;
+	int error;
 
+	error = check_if_digit_only(argc, argv[1]);
+	if (error == 1)
+		return (return_error);	
 	pid = ft_atoi(argv[1]);
 	if (argc < 3)
-	{
-		write(1, "Error", 5);
-		return (0);
-	}
+		return (return_error);
 	kill(pid, SIGUSR1);
 	printf("test:%d\n", pid);
 	printf("client\n");
